Readd index storing changes by Maarten
A merge removed the changes Maarten did to the way list indexes are
stored. They where modified to be stored as a single string instead of
multiple values on the same field.

Multiple values on the same field seemed to give wrong results for the
field length when calculating document scores. This is actually quite
important because we want to 'penalize' recipes with a bunch of extra
ingredients that were not given in the query.

Reimplementing this change makes that documents with less unmatched
tokes could end up getting betters scores.